output attempt 1 - check commit comment
will commit this milestone and continue working from here.
PLEASE DELETE THE OUTPUT FILE AFTER EVERY RUN.
For some reason, the output is gibberish if I try to open the file in overwrite mode. I think the file is overwritten even though it's opened in append mode, but just to be safe, keep the output file opened in append mode, and delete it after every run